requirements.txt
# Restaurant Simulation Project Dependencies
# Python 3.12+

# Core simulation and data science
simpy>=4.1.0
numpy>=2.0.0
pandas>=2.0.0

# Visualization
matplotlib>=3.10.0
seaborn>=0.13.0
plotly>=5.17.0

# GUI Application
streamlit>=1.28.0

# Jupyter notebook environment
jupyter>=1.0.0
ipykernel>=7.0.0

scipy>=1.10.0

# RAG Chatbot Dependencies
openai>=1.0.0
chromadb>=0.4.0
python-dotenv>=1.0.0
